adsdsd

+++ application/modules/manajemen/models/Model_users.php
@@ -20,26 +20,28 @@ class Model_users extends CI_Model
 		} else {
 			$valid = ($this->input->post('password') != '') ? 'required|' : '';
 		}
-  	$this->form_validation->set_rules('username', 'Username', 'required|trim|min_length[6]|alpha_dash');
-  	$this->form_validation->set_rules('fullname', 'Nama Lengkap', 'required|trim');
-  	$this->form_validation->set_rules('id_opd',   'Nama OPD', 'required|trim');
-  	$this->form_validation->set_rules('email',    'Email', 'trim|valid_email');
+		$this->form_validation->set_rules('username', 'Username', 'required|trim|min_length[6]|alpha_dash');
+		$this->form_validation->set_rules('fullname', 'Nama Lengkap', 'required|trim');
+		$this->form_validation->set_rules('id_opd',   'Nama OPD', 'required|trim');
+		$this->form_validation->set_rules('id_jabatan',   'Jabatan', 'required|trim');
+		$this->form_validation->set_rules('email',    'Email', 'trim|valid_email');
 		$this->form_validation->set_rules('password', 'Password', $valid.'regex_match[/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[$@$!%*?&])[A-Za-z\d$@$!%*?&]{8,}$/]');
-  	$this->form_validation->set_rules('conf_password', 'Konfirmasi Password', $valid.'matches[password]');
+		$this->form_validation->set_rules('conf_password', 'Konfirmasi Password', $valid.'matches[password]');
 		$this->form_validation->set_rules('groupid[]', 'Group User', 'required|trim');
-  	validation_message_setting();
-    if ($this->form_validation->run() == FALSE)
-      return false;
-    else
-      return true;
+		validation_message_setting();
+		if ($this->form_validation->run() == FALSE)
+			return false;
+		else
+			return true;
 	}
 
 	public function getDataListGroup()
 	{
 		$this->db->where('id_status', 1);
-		if(!$this->app_loader->is_admin()) {
+		$this->db->where('id_group !=', 1);
+		/*if(!$this->app_loader->is_admin()) {
 			$this->db->where_not_in('id_level_akses', array(1));
-		}
+		}*/
 		$this->db->order_by('id_level_akses ASC');
 		$this->db->order_by('id_group ASC');
 		$query = $this->db->get('xi_sa_group');
@@ -61,6 +63,40 @@ class Model_users extends CI_Model
 	    return $dd_opd;
    }
 
+   public function getDataListJabatan()
+   {
+	    $this->db->where('id_status', '1');
+	    $this->db->order_by('id_jabatan ASC');
+		$query = $this->db->get('xi_sa_jabatan');
+
+	    $dd_jab[''] = 'Pilih Data';
+	    if ($query->num_rows() > 0) {
+	      foreach ($query->result() as $row) {
+	        $dd_jab[$row->id_jabatan] = $row->nm_jabatan;
+	      }
+	    }
+	    return $dd_jab;
+   }
+
+   public function getDataJabatan($opdId)
+   {
+	    $this->db->where('id_status', '1');
+	    $this->db->where('id_opd', $opdId);
+	    $this->db->order_by('id_jabatan ASC');
+		$query = $this->db->get('xi_sa_jabatan');
+
+	    return $query->result_array();
+   }
+
+   public function getDetailJenisUser($id_jabatan)
+   {
+	    $this->db->where('id_status', '1');
+	    $this->db->where('id_jabatan', $id_jabatan);
+		$query = $this->db->get('xi_sa_jabatan');
+
+	    return $query->row_array()['jenis_user'];
+   }
+
    public function getDataJenisUser()
    {
 	    $this->db->where('id_status', '1');	    
@@ -77,22 +113,22 @@ class Model_users extends CI_Model
 
 	public function getDataLevelAkses()
     {
-		if(!$this->app_loader->is_admin()) {
-			$this->db->where('id_level_akses !=', 1);
-		}
-		$this->db->where('id_status', 1);
-		$query = $this->db->get('xi_sa_level_akses');
-    $dd_level[''] = 'Pilih Data';
-    if ($query->num_rows() > 0) {
-      foreach ($query->result_array() as $row) {
-        $dd_level[$row['id_level_akses']] = $row['level_akses'];
-      }
-    }
-    return $dd_level;
+    	if(!$this->app_loader->is_admin()) {
+    		$this->db->where('id_level_akses !=', 1);
+    	}
+    	$this->db->where('id_status', 1);
+    	$query = $this->db->get('xi_sa_level_akses');
+    	$dd_level[''] = 'Pilih Data';
+    	if ($query->num_rows() > 0) {
+    		foreach ($query->result_array() as $row) {
+    			$dd_level[$row['id_level_akses']] = $row['level_akses'];
+    		}
+    	}
+    	return $dd_level;
   }
 
 	var $search = array('a.username', 'a.fullname', 'c.nama_group');
-	public function get_datatables($param)
+  public function get_datatables($param)
   {
     $this->_get_datatables_query($param);
     if($_POST['length'] != -1)
@@ -101,7 +137,7 @@ class Model_users extends CI_Model
     return $query->result_array();
   }
 
-	public function count_filtered($param)
+  public function count_filtered($param)
   {
     $this->_get_datatables_query($param);
     $query = $this->db->get();
@@ -110,105 +146,105 @@ class Model_users extends CI_Model
 
   public function count_all()
   {
-		$this->db->select('a.id_users');
-		$this->db->from('xi_sa_users a');
-		$this->db->join('xi_sa_users_privileges b', 'a.id_users = b.id_users', 'INNER');
-		$this->db->join('xi_sa_group c', 'b.id_group = c.id_group', 'INNER');
-		$this->db->where('a.account_type', 1);
-		if(!$this->app_loader->is_admin()) {
-			$this->db->where('c.id_level_akses !=', 1);
-		}
-    return $this->db->count_all_results();
+  	$this->db->select('a.id_users');
+  	$this->db->from('xi_sa_users a');
+  	$this->db->join('xi_sa_users_privileges b', 'a.id_users = b.id_users', 'INNER');
+  	$this->db->join('xi_sa_group c', 'b.id_group = c.id_group', 'INNER');
+  	$this->db->where('a.account_type', 1);
+  	if(!$this->app_loader->is_admin()) {
+  		$this->db->where('c.id_level_akses !=', 1);
+  	}
+  	return $this->db->count_all_results();
   }
 
   private function _get_datatables_query($param)
   {
-		$post = array();
-		if (is_array($param)) {
-      foreach ($param as $v) {
-        $post[$v['name']] = $v['value'];
-      }
-    }
-		$this->db->select('a.id_users,
-											 a.token,
-											 a.username,
-											 a.email,
-											 a.fullname,
-											 a.foto_profile,
-											 a.blokir,
-											 a.jenis_user,
-											 a.id_status,
-											 (CASE
-											   WHEN d.pass_plain IS NULL THEN "-"
-											   ELSE d.pass_plain
-											 END) AS pass_plain,
-											 GROUP_CONCAT(c.nama_group ORDER BY c.id_level_akses ASC SEPARATOR ",") AS group_user');
-    $this->db->from('xi_sa_users a');
-		$this->db->join('xi_sa_users_privileges b', 'a.id_users = b.id_users', 'left');
-		$this->db->join('xi_sa_group c', 'b.id_group = c.id_group', 'left');
-		$this->db->join('xi_sa_users_default_pass d', 'a.id_users = d.id_users', 'left');
-		$this->db->where('b.id_status', 1);
-		$this->db->where('a.account_type', 1);
-		if(!$this->app_loader->is_admin()) {
-			$this->db->where('c.id_level_akses !=', 1);
-		}
-		//level
-		if(isset($post['level']) AND $post['level'] != '')
-			$this->db->where('c.id_level_akses', $post['level']);
-		//blokir
-		if(isset($post['blokir']) AND $post['blokir'] != '')
-			$this->db->where('a.blokir', $post['blokir']);
-		//status
-		if(isset($post['status']) AND $post['status'] != '')
-			$this->db->where('a.id_status', $post['status']);
-		//username
-		if(isset($post['username']) AND $post['username'] != '')
-			$this->db->like('a.username', $post['username'], 'after');
-		//fullname
-		if(isset($post['fullname']) AND $post['fullname'] != '')
-			$this->db->like('a.fullname', $post['fullname'], 'after');
-		//email
-		if(isset($post['email']) AND $post['email'] != '')
-			$this->db->like('a.email', $post['email'], 'after');
-		//group
-		if(isset($post['group']) AND $post['group'] != '')
-			$this->db->like('c.nama_group', $post['group'], 'after');
-    $i = 0;
-    foreach ($this->search as $item) { // loop column
-      if($_POST['search']['value']) { // if datatable send POST for search
-        if($i===0) { // first loop
-          $this->db->group_start(); // open bracket. query Where with OR clause better with bracket. because maybe can combine with other WHERE with AND.
-          $this->db->like($item, $_POST['search']['value']);
-        } else {
-          $this->db->or_like($item, $_POST['search']['value']);
-        }
+	  	$post = array();
+	  	if (is_array($param)) {
+	  		foreach ($param as $v) {
+	  			$post[$v['name']] = $v['value'];
+	  		}
+	  	}
+	  	$this->db->select('a.id_users,
+	  		a.token,
+	  		a.username,
+	  		a.email,
+	  		a.fullname,
+	  		a.foto_profile,
+	  		a.blokir,
+	  		a.id_jabatan,
+	  		a.jenis_user,
+	  		a.id_status,
+	  		(CASE
+	  			WHEN d.pass_plain IS NULL THEN "-"
+	  			ELSE d.pass_plain
+	  			END) AS pass_plain,
+	  	GROUP_CONCAT(c.nama_group ORDER BY c.id_level_akses ASC SEPARATOR ",") AS group_user');
+	  	$this->db->from('xi_sa_users a');
+	  	$this->db->join('xi_sa_users_privileges b', 'a.id_users = b.id_users', 'left');
+	  	$this->db->join('xi_sa_group c', 'b.id_group = c.id_group', 'left');
+	  	$this->db->join('xi_sa_users_default_pass d', 'a.id_users = d.id_users', 'left');
+	  	$this->db->where('b.id_status', 1);
+	  	$this->db->where('a.account_type', 1);		
+	  	$this->db->where('a.id_users !=', 1);		
+			//level
+	  	if(isset($post['level']) AND $post['level'] != '')
+	  		$this->db->where('c.id_level_akses', $post['level']);
+			//blokir
+	  	if(isset($post['blokir']) AND $post['blokir'] != '')
+	  		$this->db->where('a.blokir', $post['blokir']);
+			//status
+	  	if(isset($post['status']) AND $post['status'] != '')
+	  		$this->db->where('a.id_status', $post['status']);
+			//username
+	  	if(isset($post['username']) AND $post['username'] != '')
+	  		$this->db->like('a.username', $post['username'], 'after');
+			//fullname
+	  	if(isset($post['fullname']) AND $post['fullname'] != '')
+	  		$this->db->like('a.fullname', $post['fullname'], 'after');
+			//email
+	  	if(isset($post['email']) AND $post['email'] != '')
+	  		$this->db->like('a.email', $post['email'], 'after');
+			//group
+	  	if(isset($post['group']) AND $post['group'] != '')
+	  		$this->db->like('c.nama_group', $post['group'], 'after');
+	  	$i = 0;
+	    foreach ($this->search as $item) { // loop column
+	      if($_POST['search']['value']) { // if datatable send POST for search
+	        if($i===0) { // first loop
+	          $this->db->group_start(); // open bracket. query Where with OR clause better with bracket. because maybe can combine with other WHERE with AND.
+	          $this->db->like($item, $_POST['search']['value']);
+	      } else {
+	      	$this->db->or_like($item, $_POST['search']['value']);
+	      }
 
-        if(count($this->search) - 1 == $i) //last loop
-          $this->db->group_end(); //close bracket
-      }
-      $i++;
-    }
-		$this->db->group_by('a.id_users');
-		$this->db->order_by('a.id_users ASC');
+	        if(count($this->search) - 1 == $i) //last loop
+	          $this->db->group_end(); //close bracket
+	      }
+	      $i++;
+	  }
+	  $this->db->group_by('a.id_users');
+	  $this->db->order_by('a.id_users ASC');
   }
 
 	/*Fungsi get data edit by id dan url*/
 	public function getDataDetailUsers($token)
 	{
 		$this->db->select('a.id_users,
-											 a.token,
-											 a.username,
-											 a.password,
-											 a.email,
-											 a.fullname,
-											 a.foto_profile,
-											 a.id_opd,
-											 a.sub_opd,
-											 a.nm_sub_opd,
-											 a.jenis_user,
-											 a.blokir,
-											 a.id_status,
-											 GROUP_CONCAT(b.id_group ORDER BY b.id_group ASC SEPARATOR ",") AS group_user');
+			a.token,
+			a.username,
+			a.password,
+			a.email,
+			a.fullname,
+			a.foto_profile,
+			a.id_opd,
+			a.sub_opd,
+			a.nm_sub_opd,
+			a.id_jabatan,
+			a.jenis_user,
+			a.blokir,
+			a.id_status,
+			GROUP_CONCAT(b.id_group ORDER BY b.id_group ASC SEPARATOR ",") AS group_user');
 		$this->db->from('xi_sa_users a');
 		$this->db->join('xi_sa_users_privileges b', 'a.id_users = b.id_users', 'left');
 		$this->db->where('a.token', $token);
@@ -216,8 +252,8 @@ class Model_users extends CI_Model
 		$this->db->group_by('a.id_users');
 		$this->db->order_by('a.id_users ASC');
 		$this->db->limit(1);
-    $query = $this->db->get();
-    return $query->row_array();
+		$query = $this->db->get();
+		return $query->row_array();
 	}
 
 	private function getDataUserByToken($token)
@@ -239,6 +275,7 @@ class Model_users extends CI_Model
 		$group			 = escape($this->input->post('groupid', TRUE));
 		$fullname    = escape($this->input->post('fullname', TRUE));
 		$id_opd    = escape($this->input->post('id_opd', TRUE));
+		$id_jabatan    = escape($this->input->post('id_jabatan', TRUE));
 		$nm_opd    = escape($this->getNamaInstansi($id_opd));
 
 		$sub_opd    = escape($this->input->post('sub_opd', TRUE));
@@ -258,7 +295,8 @@ class Model_users extends CI_Model
 			'nm_opd' 				=> $nm_opd,
 			'sub_opd' 				=> $sub_opd,
 			'nm_sub_opd' 			=> $nm_sub_opd,
-			'jenis_user' 			=> escape($this->input->post('jenis_user', TRUE)),
+			'id_jabatan' 			=> $id_jabatan,
+			'jenis_user' 			=> escape($this->getDetailJenisUser($id_jabatan)),
 			'validate_email_code'	=> '',
 			'validate_email_status'	=> 0,
 			'reset_password_code'	=> '',
@@ -302,6 +340,7 @@ class Model_users extends CI_Model
 		$password 	= escape($this->input->post('password', TRUE));
 		$group		= escape($this->input->post('groupid', TRUE));
 		
+		$id_jabatan     = escape($this->input->post('id_jabatan', TRUE));
 		$id_opd     = escape($this->input->post('id_opd', TRUE));
 		$nm_opd     = escape($this->getNamaInstansi($id_opd));
 
@@ -318,7 +357,8 @@ class Model_users extends CI_Model
 			'nm_opd' 		=> $nm_opd,
 			'sub_opd' 		=> $sub_opd,
 			'nm_sub_opd' 	=> $nm_sub_opd,
-			'jenis_user' 	=> escape($this->input->post('jenis_user', TRUE)),
+			'id_jabatan' 	=> $id_jabatan,
+			'jenis_user'    => escape($this->getDetailJenisUser($id_jabatan)),
 			'mod_by' 	  	=> $create_by,
 			'mod_date'  	=> $create_date,
 			'mod_ip'    	=> $create_ip
@@ -375,15 +415,31 @@ class Model_users extends CI_Model
 	public function deleteDataUsers()
 	{
 		$users = escape($this->input->post('tokenId', TRUE));
-		//jika ingin menghapus data lakukan looping
-		foreach ($users as $id) {
-			/*query delete*/
-			$this->db->where('token', $id);
-			$this->db->delete('xi_sa_users');
+		
+		if ($this->countAgenda($users) > 0) 
+		{
+			return array('message'=>'ERROR');
 		}
-		return array('message'=>'SUCCESS');
-	}
+		else
+		{
+			foreach ($users as $id) 
+			{
+				$this->db->where('token', $id);
+				$this->db->delete('xi_sa_users');
+				return array('message'=>'SUCCESS');
+			}
+		}
 
+		
+		
+	}
+    
+    public function countAgenda($users)
+    {
+    	$this->db->where_in('penerima', $users);
+		return $this->db->count_all_results('data_agenda');	
+
+    }
 	public function searchDataUsername($username)
   {
 		$this->db->where('username', $username);
@@ -419,20 +475,19 @@ class Model_users extends CI_Model
 	}
 
 	public function getInstansiAll()
-	{     
+	{    
+		if(!$this->app_loader->is_admin()) 
+		{
+			$id_opd = $this->app_loader->current_opd();
+			$child_ins   = $id_opd;
+		} 
+		else
+		{
+			$id_opd      = escape($this->input->get('opdId', TRUE));
+			$child_ins   = 1;  
+		}
 
-	        if(!$this->app_loader->is_admin()) 
-	        {
-	          $id_opd = $this->app_loader->current_opd();
-	          $child_ins   = $id_opd;
-	        } 
-	        else
-	        {
-	          $id_opd      = escape($this->input->get('opdId', TRUE));
-	          $child_ins   = 1;  
-	        }
-	        
-	        return $this->getInstansiParent($id_opd, $child_ins);
+		return $this->getInstansiParent($id_opd, $child_ins);
 	}
   
     public function getInstansiParent($id_opd, $child_ins)
